requirements.txt
django-hashid-field
